requirements.txt
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
cachetools==5.4.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
google==3.0.0
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
grpcio==1.65.1
h5py==3.10.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.2
ijson==3.3.0
importlib_metadata==8.2.0
litellm==1.40.29
MarkupSafe==2.1.5
mattsollamatools==0.0.25
mkl-service
multidict==6.0.5
ollama==0.1.9
openai==1.37.1
protobuf
pulsejet==0.1.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dotenv==1.0.1
python-magic==0.4.27
regex==2024.7.24
rsa==4.9
sniffio==1.3.1
tiktoken==0.7.0
typing_extensions==4.12.2
uritemplate==4.1.1
yarl==1.9.4
zipp==3.19.2
nltk~=3.8.1
numpy~=1.26.4
tqdm~=4.66.2
requests~=2.31.0
beautifulsoup4~=4.12.3
pyyaml~=6.0.1